requirements.txt
duolingo-api==0.5.0
requests==2.24.0
idna==2.9
urllib3==1.25.9
certifi==2020.6.20
chardet==3.0.4
